refactor: split common components
